requirements.txt
neo4j>=5.19.0
python-dotenv>=1.0.0
jupyter>=1.0.0
pandas>=2.0.0
requests>=2.31.0
anthropic>=0.49.0
typing-extensions>=4.0.0
mcp[cli]>=1.0.0
